requirements.txt
# Core dependencies
openai>=1.0.0
python-dotenv>=1.0.0

# Vector store
chromadb>=0.4.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20

# Document processing
python-docx>=1.0.0
pypdf2>=3.0.0
python-pptx>=0.6.0

# Google Drive (prototype)
google-api-python-client>=2.0.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.0.0

# SharePoint (production)
Office365-REST-Python-Client>=2.5.0

# Web interface
streamlit>=1.28.0

# Utilities
requests>=2.31.0
tenacity>=8.2.0

# Testing dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
pytest-timeout>=2.2.0
pytest-xdist>=3.5.0

